Handle no endpoint methods defined
