+++ src/api/base.spec.ts
@@ -1,63 +1,116 @@
 import { HttpMethod } from "@aws-cdk/aws-apigatewayv2"
-import { getApiViewMetadata, getFunctionMetadata, getSubRouteMetadata } from "../metadata"
-import { AlbumApi, topSongsFuncInner, topSongsHandler } from "../test/sampleApp"
+import { ApiEvent, ApiViewBase } from "./base"
+import { ApiView, SubRoute } from "../registry"
+import { AlbumApi } from "../test/sampleApp"
 
-describe("Metadata decorators", () => {
-  describe("@ApiView decorator", () => {
-    it("stores metadata", () => {
-      expect(getApiViewMetadata(AlbumApi)).toMatchObject({
-        apiClass: AlbumApi,
-        entry: /sample-app.ts$/,
-        memorySize: 512,
+interface IMakeApiEvent {
+  method: HttpMethod
+  path?: string
+  routeKey: string
+}
+function makeApiEvent({ method, routeKey, path = "/" }: IMakeApiEvent): Partial<ApiEvent> {
+  return {
+    requestContext: {
+      accountId: "foo",
+      apiId: "bar",
+      domainName: "baz",
+      domainPrefix: "blargle",
+      http: {
+        method,
+        path,
+        protocol: "https",
+        sourceIp: "127.0.0.1",
+        userAgent: "Konqueror",
+      },
+      requestId: "a",
+      routeKey: `${method.toUpperCase()} ${routeKey}`,
+      stage: "prod",
+      time: "abc",
+      timeEpoch: 123,
+    },
+    routeKey: `${method.toUpperCase()} ${routeKey}`,
+  }
+}
+
+describe("ApiViewBase", () => {
+  const view = new AlbumApi()
+
+  describe("findHandler", () => {
+    it("locates handler method based on request verb", () => {
+      const request = makeApiEvent({
+        method: HttpMethod.POST,
         path: "/album",
-        environment: {
-          LOG_LEVEL: "DEBUG",
-        },
+        routeKey: "/album",
       })
-    })
-  })
 
-  describe("@SubRoute decorator", () => {
-    it("stores metadata for subroutes", () => {
-      // get meta
-      const methodMeta = getSubRouteMetadata(AlbumApi)
-      const expectedMeta = {
-        propertyKey: "like",
-        requestHandlerFunc: AlbumApi.prototype.like,
-        path: "/{albumId}/like",
-        methods: [HttpMethod.POST, HttpMethod.DELETE],
+      expect(view.findHandler(request as ApiEvent)).toEqual(view.post)
+    })
+
+    it("locates appropriate method based on request", () => {
+      const request = makeApiEvent({
+        method: HttpMethod.POST,
+        path: "/album/123/like",
+        routeKey: "/album/{albumId}/like",
+      })
+
+      expect(view.findHandler(request as ApiEvent)).toEqual(view.like)
+    })
+
+    it("locates appropriate method with ANY verb", () => {
+      @ApiView({
+        path: "/proxy",
+      })
+      class AnyApi extends ApiViewBase {
+        @SubRoute({ path: "/sub", methods: [HttpMethod.ANY] })
+        async handler() {
+          return "OK"
+        }
       }
-      expect(methodMeta.size).toBe(1)
-      expect(methodMeta.get("like")).toMatchObject(expectedMeta)
-    })
-  })
 
-  describe("Function()", () => {
-    it("stores metadata on functions with separate Lambda() call", () => {
-      const funcMeta = getFunctionMetadata(topSongsHandler)
-      expect(funcMeta).toMatchObject({
-        entry: /test\/sampleApp.ts$/,
-        handler: "topSongsHandler",
-        requestHandlerFunc: topSongsHandler,
-        path: "/top-songs",
-        methods: [HttpMethod.PUT],
-        environment: {
-          LOG_LEVEL: "WARN",
-        },
+      const request = makeApiEvent({
+        method: HttpMethod.POST,
+        path: "/proxy/sub",
+        routeKey: "/proxy/sub",
       })
+
+      const anyView = new AnyApi()
+      expect(anyView.findHandler(request as ApiEvent)).toEqual(anyView.handler)
     })
-    it("stores metadata on functions wrapped with Lambda()", () => {
-      const funcMeta = getFunctionMetadata(topSongsFuncInner)
-      expect(funcMeta).toMatchObject({
-        entry: /test\/sampleApp.ts$/,
-        handler: "topSongsFuncInner",
-        requestHandlerFunc: topSongsFuncInner,
-        path: "/top-songs-inner",
-        methods: [HttpMethod.PUT],
-        environment: {
-          LOG_LEVEL: "WARN",
-        },
+
+    it("locates appropriate class route with any method", () => {
+      @ApiView({
+        path: "/proxy",
       })
+      class AnyApi extends ApiViewBase {}
+
+      const request = makeApiEvent({
+        method: HttpMethod.POST,
+        path: "/proxy",
+        routeKey: "/proxy",
+      })
+
+      const anyView = new AnyApi()
+      expect(anyView.findHandler(request as ApiEvent)).toEqual(anyView.post)
+    })
+
+    it("locates no method based on request if no match on routeKey", () => {
+      const request = makeApiEvent({
+        method: HttpMethod.POST,
+        path: "/album/123/like",
+        routeKey: "/album/x/{albumId}/like",
+      })
+
+      expect(view.findHandler(request as ApiEvent)).toBeUndefined()
+    })
+
+    it("locates no method based on request if no match on method", () => {
+      const request = makeApiEvent({
+        method: HttpMethod.HEAD,
+        path: "/album/123/like",
+        routeKey: "/album/{albumId}/like",
+      })
+
+      expect(view.findHandler(request as ApiEvent)).toBeUndefined()
     })
   })
 })
